Now correctly removing presenter in onDestroy rather than onStop so can now go back to correctly creating and initialising the presenter in onCreate rather than in onResume.
When recreating the activity after an orientation/configuration change onItemSelected is re-called.  The spinner selection won't have changed so we don't bother fetching a new movie list.

Because we are now using RX Java to load the movies in the background onResume will be called before we have finished loading the movies.  The recyclerview state will have already been restored in onResume, we then override that by setting the recyclerview up again after a configuration change.  We therefore have to restore the recyclerview state after we have items to maintain the scroll position across orientation changes.